# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MARSA Yacht Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist/{yachtId}:
    parameters:
      - name: yachtId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: addWishlistItem
      tags: [wishlist]
      summary: Add a live yacht to the wishlist
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistMutationResponse"
    delete:
      operationId: removeWishlistItem
      tags: [wishlist]
      summary: Remove a yacht from the wishlist
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistMutationResponse"
...
components:
  schemas:
...
    WishlistMutationResponse:
      type: object
      properties:
        yachtId: { type: string }
        wishlisted: { type: boolean }
      required: [yachtId, wishlisted]
...